Added logic to posts where only user can access their posts.

+++ app/routers/post.py
@@ -10,26 +10,25 @@ router = APIRouter(prefix="/posts", tags=["Posts"])
 
 @router.get("", response_model=List[schemas.Post])
 def get_posts(db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
-    print(current_user.id)
-    posts = db.query(models.Post).order_by(models.Post.id.asc()).all()
+    posts = db.query(models.Post).all()
     return posts
 
 
 @router.get("/{id}", response_model=schemas.Post)
 def get_post(id: int, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
-    print(current_user)
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id {id} not found")
+    # if post.owner_id != current_user.id:
+    #     raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform this operation!")
     return post
 
 
 @router.post("", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
 def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db),
                  current_user=Depends(oauth2.get_current_user)):
-    print(current_user)
-    new_post = models.Post(**post.dict())
+    new_post = models.Post(owner_id=current_user.id, **post.dict())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -38,13 +37,13 @@ def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db),
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
 def delete_post(id: int, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
-    print(current_user)
-
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id {id} does not exist")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform this operation!")
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -53,13 +52,13 @@ def delete_post(id: int, db: Session = Depends(get_db), current_user=Depends(oau
 @router.put("/{id}", response_model=schemas.Post)
 def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends(get_db),
                 current_user=Depends(oauth2.get_current_user)):
-    print(current_user)
-
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     if post is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id {id} does not exist")
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform this operation!")
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
     return post_query.first()
